Add basic result ranking (BM_25)
Oh, damn this was a lot of work (and triel and error) to get here
almost the whole index changed, but I now know better than ever that
the index in the current form does not hold up to larger datasets.

So sharding will be definitly necessary to scale to million webpages.
Well, I kinda knew that but even 10_000 is currently impossible which
just sucks.